pyproject.toml
[project]
name = "nhp_data"
version = "0.1.0"
description = "Add your description here"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.13.4",
  "databricks-connect>=15.4",
  "pandas>=2.2.3",
  "openpyxl>=3.1.5",
  "pygam>=0.9.1",
  "pyjanitor>=0.31.0",
  "pyspark>=3.5.1",
  "tqdm>=4.67.1",
  "xlrd>=2.0.2"
]

# Explicitly specify which packages to include
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nhp.data.raw_data.mitigators.reference_data" = ["*.json", "*.csv"]

[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
aggregated_data-ecds = "nhp.data.aggregated_data.ecds:main"
aggregated_data-outpatients = "nhp.data.aggregated_data.outpatients:main"

default-apc = "nhp.data.default.apc:main"
default-ecds = "nhp.data.default.ecds:main"
default-opa = "nhp.data.default.opa:main"

inputs_data-catchments = "nhp.data.inputs_data.catchments:main"
inputs_data-age_sex = "nhp.data.inputs_data.age_sex:main"
inputs_data-covid_adjustment = "nhp.data.inputs_data.covid_adjustment:main"
inputs_data-diagnoses = "nhp.data.inputs_data.diagnoses:main"
inputs_data-expat_repat = "nhp.data.inputs_data.expat_repat:main"
inputs_data-procedures = "nhp.data.inputs_data.procedures:main"
inputs_data-rates = "nhp.data.inputs_data.rates:main"
inputs_data-wli = "nhp.data.inputs_data.wli:main"
inputs_data-inequalities = "nhp.data.inputs_data.inequalities:main"

raw_data-aae = "nhp.data.raw_data.aae:main"
raw_data-ecds = "nhp.data.raw_data.ecds:main"
raw_data-inpatients = "nhp.data.raw_data.inpatients:main"
raw_data-inpatients_mitigators = "nhp.data.raw_data.inpatients_mitigators:generate_inpatients_mitigators"
raw_data-outpatients = "nhp.data.raw_data.outpatients:main"
raw_data-outpatients_mitigators = "nhp.data.raw_data.outpatients_mitigators:main"

population_projections-get_ons_files_2022 = "nhp.data.population_projections.get_ons_files_2022:main"
population_projections-npp = "nhp.data.population_projections.npp:main"
population_projections-snpp = "nhp.data.population_projections.snpp:main"

model_data-health_status_adjustment-generate_gams = "nhp.data.model_data.health_status_adjustment.generate_gams:main"
model_data-health_status_adjustment-generate_national_gams = "nhp.data.model_data.health_status_adjustment.generate_national_gams:main"
model_data-aae = "nhp.data.model_data.aae:main"
model_data-birth_factors = "nhp.data.model_data.birth_factors:main"
model_data-demographic_factors = "nhp.data.model_data.demographic_factors:main"
model_data-ip = "nhp.data.model_data.ip:main"
model_data-op = "nhp.data.model_data.op:main"

reference-ods_trusts = "nhp.data.reference.ods_trusts:main"
reference-provider_main_icb = "nhp.data.reference.provider_main_icb:generate_provider_main_icb"
reference-population_by_imd_decile = "nhp.data.reference.population_by_imd_decile:main"
reference-provider_catchments = "nhp.data.reference.provider_catchments:main"
